requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
pydantic>=2.6
SQLAlchemy>=2.0
psycopg2-binary>=2.9
faiss-cpu>=1.8
mlflow>=2.13
evidently>=0.4.27
langchain>=0.2
llama-index>=0.11
pytest>=8.2
